♻️ refactor: UserProfilePage와 UserProfilePageClient로 분리
